Move default user setting to API
