+++ src/api/user/index.ts
@@ -1,2 +1,3 @@
 export * from './types';
 export * from './user';
+export * from './models';
