+++ src/pages/search/exportcitation/[format].tsx
@@ -1,4 +1,5 @@
 import {
+  DEFAULT_USER_DATA,
   ExportApiFormatKey,
   exportCitationKeys,
   fetchExportCitation,
@@ -11,7 +12,7 @@ import {
 } from '@api';
 import { Alert, AlertIcon, Box, Flex, Heading, HStack, Link } from '@chakra-ui/react';
 import { ChevronLeftIcon } from '@chakra-ui/icons';
-import { CitationExporter, DEFAULT_USER_DATA, JournalFormatMap } from '@components';
+import { CitationExporter, JournalFormatMap } from '@components';
 import { getExportCitationDefaultContext } from '@components/CitationExporter/CitationExporter.machine';
 import { APP_DEFAULTS } from '@config';
 import { useIsClient } from '@lib/useIsClient';
